// src/app/models/admin/hosogiayto/tailieu.model.ts
export class OutputHsTaiLieuModel {
  idtailieu: string;
  tentailieu: string;
  filedinhkem: string;
  duongdan: string;
  sobanchinh: number;
  sobansao: number;
  idloaitailieu: string;
  tenloaitailieu: string;
  idhoso: string;
  dungluong: number;
  dinhdang: string;
  nhomtailieu: number;
  thutu: number;
  siteid: string;
}

export class InputHsTaiLieuModel {
  idtailieu: string;
  tentailieu: string;
  filedinhkem: string;
  duongdan: string;
  sobanchinh: number;
  sobansao: number;
  idloaitailieu: string;
  tenloaitailieu: string;
  idhoso: string;
  dungluong: number;
  dinhdang: string;
  nhomtailieu: number;
  thutu: number;
  siteid: string;
}